Components reorganized
