+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
-import Navegation from "./components/Navegation.js";
-import ListPublication from "./components/ListPublication.js";
-import Footer from "./components/Footer.js";
+import Navigation from "./components/Navigation/Navigation";
+import ListPublication from "./components/Publication/ListPublication";
+import Footer from "./components/Footer/Footer";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import StoryPage from "./components/StoryPage";
+import StoryPage from "./components/Publication/StoryPage";
 
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
 import muiTheme from './theme/muiTheme';
@@ -13,11 +13,11 @@ import muiTheme from './theme/muiTheme';
 class App extends Component {
   render() {
     return (
-      
+
       <MuiThemeProvider theme={muiTheme}>
        <Router>
         <div>
-          <Navegation />
+          <Navigation />
           <Route exact path="/" component={ListPublication} />
           <Route path="/stories/:id" component={StoryPage} />
           <Footer />
@@ -28,6 +28,3 @@ class App extends Component {
   }
 }
 export default App;
-
-
-
