+++ src/components/Navigation/Navigation.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react'
 import AppBar from '@material-ui/core/AppBar'
 import Toolbar from '@material-ui/core/Toolbar'
 import { Link } from "react-router-dom";
-import logo from './logo.png';
+import logo from '../../Images/nearshorelogo.png';
 
-class Navegation extends Component {
+class Navigation extends Component {
     render() {
     return (
         <div>
@@ -18,4 +18,4 @@ class Navegation extends Component {
 }
 }
 
-export default Navegation;
+export default Navigation;
